requirements.txt
langchain==0.3.7
langchain-openai==0.2.8
langgraph==0.2.50
python-dotenv==1.0.1